requirements.txt
pandas==3.0.1
SQLAlchemy==2.0.48
psycopg2-binary==2.9.11
pyarrow==23.0.1
openpyxl==3.1.5
matplotlib==3.10.8
numpy==2.4.3
python-dotenv==1.1.1
